# schema/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Infrahub",
        "version": "latest"
    },
    "paths": {
        "/api/artifact/{artifact_id}": {
            "get": {
                "summary": "Get Artifact",
                "operationId": "get_artifact_api_artifact__artifact_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    },
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "artifact_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Artifact Id"
                        }
                    },
                    {
                        "name": "branch",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Name of the branch to use for the query",
                            "title": "Branch"
                        },
                        "description": "Name of the branch to use for the query"
                    },
                    {
                        "name": "at",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Time to use for the query, in absolute or relative format",
                            "title": "At"
                        },
                        "description": "Time to use for the query, in absolute or relative format"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/artifact/generate/{artifact_definition_id}": {
            "post": {
                "summary": "Generate Artifact",
                "operationId": "generate_artifact_api_artifact_generate__artifact_definition_id__post",
                "security": [
                    {
                        "HTTPBearer": []
                    },
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "artifact_definition_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Artifact Definition Id"
                        }
                    },
                    {
                        "name": "branch",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Name of the branch to use for the query",
                            "title": "Branch"
                        },
                        "description": "Name of the branch to use for the query"
                    },
                    {
                        "name": "at",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Time to use for the query, in absolute or relative format",
                            "title": "At"
                        },
                        "description": "Time to use for the query, in absolute or relative format"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArtifactGeneratePayload",
                                "description": "Payload of the request, can be used to limit the scope of the query to a specific list of hosts",
                                "default": {
                                    "nodes": []
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ArtifactGeneratePayload": {
                "properties": {
                    "nodes": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Nodes"
                    }
                },
                "type": "object",
                "title": "ArtifactGeneratePayload"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            },
            "APIKeyHeader": {
                "type": "apiKey",
                "in": "header",
                "name": "X-INFRAHUB-KEY"
            }
        }
    }
}